Refactor module and import paths for user, roles, and posts modules
- rename `src/users/`, `src/roles/`, `src/posts/` to `src/modules/users/`, `src/
