+++ src/app.module.ts
@@ -1,14 +1,14 @@
 import { Module } from '@nestjs/common'
 import { SequelizeModule } from '@nestjs/sequelize'
-import { UsersModule } from './users/users.module'
+import { UsersModule } from './modules/users/users.module'
 import { ConfigModule } from '@nestjs/config'
-import { User } from './users/model/users.model'
-import { RolesModule } from './roles/roles.module'
-import { Role } from './roles/model/roles.model'
-import { UserRoles } from './roles/model/user-roles.model'
+import { User } from './modules/users/model/users.model'
+import { RolesModule } from './modules/roles/roles.module'
+import { Role } from './modules/roles/model/roles.model'
+import { UserRoles } from './modules/roles/model/user-roles.model'
 import { AuthModule } from './modules/auth/auth.module'
-import { PostsModule } from './posts/posts.module'
-import { Post } from './posts/model/posts.model'
+import { PostsModule } from './modules/posts/posts.module'
+import { Post } from './modules/posts/model/posts.model'
 import { FilesModule } from './modules/files/files.module'
 import { ServeStaticModule } from '@nestjs/serve-static'
 import * as path from 'path'
